+++ cextern/trim_wcslib.sh
@@ -11,3 +11,4 @@ rm -rf wcslib/html
 rm -rf wcslib/pgsbox
 rm -rf wcslib/utils
 rm wcslib/*.pdf
+patch -p1 < patches/unit-translate-information.diff